benchmarks/python/thumbnail/requirements.txt
minio
urllib3
Pillow